refactor: components-event 디렉터리 생성 및 이벤트 파일 이동
