+++ index.html
@@ -55,6 +55,5 @@
           
     </div>
     <script type="module" src="./src/main.ts"></script>
-    <script type="module" src="./src/utils/ClickEvent.js"></script>
   </body>
 </html>
