+++ src/main.ts
@@ -1,8 +1,9 @@
 import { FilterBox } from "./components/common/FilterBox.js";
 import { LunchList } from "./components/function/LunchList.ts";
 import { RestaurantTabMenu } from "./components/function/RestaurantTabMenu.js";
-import ChangeEvent from "./utils/ChangeEvent.ts";
-import SubmitEvent from "./utils/SubmitEvent.ts";
+import ChangeEvent from "./components-event/ChangeEvent.ts";
+import SubmitEvent from "./components-event/SubmitEvent.ts";
+import ClickEvent from "./components-event/ClickEvent.js";
 
 RestaurantTabMenu("restaurantTabMenuBox");
 const filterBox = FilterBox();
@@ -12,3 +13,4 @@ lunchList.render();
 lunchList.renderFavorites();
 ChangeEvent(lunchList);
 SubmitEvent(lunchList);
+new ClickEvent(document);
